+++ pkg/security/secl/model/model_unix.go
@@ -529,7 +529,7 @@ type Mount struct {
 	Origin               uint32   `field:"-"`
 	Detached             bool     `field:"detached"` // SECLDoc[detached] Definition:`Mount is detached from the VFS`
 	Visible              bool     `field:"visible"`  // SECLDoc[visible] Definition:`Mount is not visible in the VFS`
-	NamespaceInode       uint32
+	NamespaceInode       uint64
 }
 
 // MountEvent represents a mount event
